pyproject.toml
[project]
name = "kqlalchemy"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Kyllo", email = "alex.kyllo@gmail.com"},
]
dependencies = [
    "sqlalchemy~=1.4",
    "azure-identity",
    "pandas",
    "pyodbc",
    "azure-kusto-data>=2.3.2",
]
requires-python = ">=3.7.1"
license = {text = "MIT"}

[project.urls]
homepage = ""

[project.optional-dependencies]
[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "sphinx",
    "black>=21.12b0",
    "pylint>=2.12.2",
    "isort>=5.10.1",
    "pytest>=6.2.5",
]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
